get all users class and routing
