+++ views/__init__.py
@@ -1,0 +1,1 @@
+from .user_request import get_all_users
